refactor(structures) create structures dir
